install custom resources, patch the pod CIDR. Calico running, but in degraded state.

+++ ubuntu-utm-vms/vm-prep/install-calico-cni.sh
@@ -1,4 +1,13 @@
 #!/bin/bash
+set -e -x
+
+KUBECONFIG=/usr/share/host/guest/generated/users/calico-cni.kubeconfig
+if  [ ! -f ${KUBECONFIG} ] ; then
+    echo "Kubeconfig for Calico is missing! $KUBECONFIG"
+    exit 1;
+fi
+cp ${KUBECONFIG} /etc/cni/net.d/calico-kubeconfig
+chmod 600 /etc/cni/net.d/calico-kubeconfig
 
 cp /usr/share/host/guest/all-nodes/downloads/calico /opt/cni/bin/
 cp /usr/share/host/guest/all-nodes/downloads/calico-ipam /opt/cni/bin/
